CRAB-44303: consolidate agent otp processes in the helper file
