+++ Seeq.Link.SDK.Debugging.Agent/AgentKeyHelper.cs
@@ -1,4 +1,7 @@
 ﻿using System.IO;
+using Seeq.Link.SDK.Services;
+using Seeq.Link.SDK.Utilities;
+using Seeq.Utilities;
 
 namespace Seeq.Link.Debugging.Agent {
 
@@ -6,12 +9,29 @@ namespace Seeq.Link.Debugging.Agent {
         private const string AGENT_ONE_TIME_PASSWORD_PLACEHOLDER = "<your_agent_one_time_password>";
         private static string OtpFilePath = Path.Combine("data", "keys", "agent.otp");
 
-        public static bool IsAgentOneTimePasswordSet() {
-            var fileContent = ReadAgentOneTimePassword();
+        public static void SetupAgentOtp(string seeqDataFolder, string agentName) {
+            if (isAgentOneTimePasswordSet()) {
+                var agentHelper = new AgentHelper(agentName);
+                var secretsPath = Path.Combine(seeqDataFolder, SeeqNames.Agents.AgentKeysFolderName, "agent.keys");
+                var secretsManager = new FileBasedSecretsManager(secretsPath);
+
+                // set the agent's pre-provisioned one-time password
+                var agentOneTimePassword = readAgentOneTimePassword();
+                var preProvisionedOneTimePasswordSecretName =
+                    $"{agentHelper.ProvisionedAgentUsername}|PRE_PROVISIONED_ONE_TIME_PASSWORD";
+                secretsManager.PutSecret(preProvisionedOneTimePasswordSecretName, agentOneTimePassword);
+
+                // clear the OTP
+                resetAgentOneTimePasswordFile();
+            }
+        }
+
+        private static bool isAgentOneTimePasswordSet() {
+            var fileContent = readAgentOneTimePassword();
             return fileContent != AGENT_ONE_TIME_PASSWORD_PLACEHOLDER;
         }
 
-        public static string ReadAgentOneTimePassword() {
+        private static string readAgentOneTimePassword() {
             if (!File.Exists(OtpFilePath)) {
                 return null;
             }
@@ -19,7 +39,7 @@ namespace Seeq.Link.Debugging.Agent {
             return File.ReadAllText(OtpFilePath).Trim();
         }
 
-        public static void ResetAgentOneTimePasswordFile() {
+        private static void resetAgentOneTimePasswordFile() {
             File.WriteAllText(OtpFilePath, AGENT_ONE_TIME_PASSWORD_PLACEHOLDER);
         }
     }
